requirements.txt
APScheduler==3.11.1
babel==2.17.0
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
cryptography==46.0.3
cryptography-fernet==0.1.0
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
Faker==37.12.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.11
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
libpass==1.9.3
MarkupSafe==3.0.3
pycparser==2.23
python-dotenv==1.2.1
PyYAML==6.0.3
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1